Restructure managers to divide responsibilities on kinds of tasks
