+++ internal/overlord/workspacestate/handlers.go
@@ -1,24 +1,16 @@
-package workspace
+package workspacestate
 
 import (
-	"encoding/json"
 	"fmt"
-	"path/filepath"
-	"strconv"
 
-	util "github.com/canonical/workspace/internal"
-	store "github.com/canonical/workspace/internal/fakestore"
-	"github.com/canonical/workspace/internal/logger"
 	"github.com/canonical/workspace/internal/overlord/projectstate"
+	. "github.com/canonical/workspace/internal/overlord/sharedstate"
+
 	"github.com/canonical/workspace/internal/overlord/state"
-	srv "github.com/canonical/workspace/internal/workspacebackend"
+	backend "github.com/canonical/workspace/internal/workspacebackend"
+
 	"gopkg.in/tomb.v2"
 )
-
-type SdkSequenceRecord struct {
-	Channel  string `json:"channel"`
-	Revision int64  `json:"revision"`
-}
 
 func (m *WorkspaceManager) undoCreateWorkspace(task *state.Task, tomb *tomb.Tomb) error {
 	project, workspace, err := ProjectAndWorkspace(task)
@@ -63,7 +55,7 @@ func (m *WorkspaceManager) doMountProject(task *state.Task, tomb *tomb.Tomb) err
 	}
 
 	/* Configure workspace core properties: project directory */
-	var prjMount = srv.WorkspaceDevice{
+	var prjMount = backend.WorkspaceDevice{
 		Name:       projectstate.ProjectDevice,
 		Properties: map[string]string{"type": "disk", "source": project.Path, "path": "/project"},
 	}
@@ -113,307 +105,4 @@ func (m *WorkspaceManager) undoStart(task *state.Task, tomb *tomb.Tomb) error {
 	}
 
 	return nil
-}
-
-func (m *WorkspaceManager) doInstallSDK(task *state.Task, tomb *tomb.Tomb) error {
-	project, workspace, err := ProjectAndWorkspace(task)
-	if err != nil {
-		return err
-	}
-
-	blob, err := sdkData(task)
-	if err != nil {
-		return err
-	}
-
-	st := task.State()
-	st.Lock()
-	defer st.Unlock()
-
-	fmt.Printf("Setting up SDK \"%s\" from %s revision %d...\n", blob.Name, blob.Channel, blob.Revision)
-
-	sdkMount := sdkBlobDevice(blob)
-
-	err = m.backend.AddWorkspaceDevice(workspace, project.ProjectId, sdkMount)
-	if err != nil {
-		return err
-	}
-
-	cleanup := func() {
-		/* Make sure the SDK file will be unmounted once installed into the workspace */
-		if err := m.backend.RemoveWorkspaceDevice(workspace, project.ProjectId, sdkMount.Name); err != nil {
-			logger.Debugf("cannot unmount SDK blob %q from workspace %q: %v", sdkMount.Name, workspace, err)
-		}
-	}
-
-	defer cleanup()
-
-	/* example: /var/lib/workspace/sdk/cuda/712/ */
-	sdkPath := filepath.Join(util.WorkspaceSdksDir, blob.Name,
-		strconv.Itoa(int(blob.Revision)))
-
-	/* Unpack the SDK to the desired location in the workspace
-	   Note: the following command requires ~ tar >= 1.29 due to --one-top-level */
-	args := srv.ExecArgs{
-		User: "root",
-		Command: []string{
-			"tar",
-			"--extract",
-			"--file",
-			sdkMount.Properties["path"],
-			"--one-top-level=" + sdkPath,
-			"--no-same-owner",
-			"--strip-components=1",
-		},
-		WorkDir: "/",
-		Stdin:   nil,
-		Stdout:  nil,
-		Stderr:  nil}
-	done, err := m.backend.Exec(workspace, project.ProjectId, &args)
-
-	/* The server will close this channel when exec is finished and no i/o remains outstanding */
-	<-done
-
-	return err
-}
-
-func (m *WorkspaceManager) undoInstallSdk(task *state.Task, tomb *tomb.Tomb) error {
-	project, workspace, err := ProjectAndWorkspace(task)
-	if err != nil {
-		return err
-	}
-
-	blob, err := sdkData(task)
-	if err != nil {
-		return err
-	}
-
-	st := task.State()
-	st.Lock()
-	defer st.Unlock()
-	sdkMount := sdkBlobDevice(blob)
-
-	fs, err := m.backend.GetWorkspaceFs(workspace, project.ProjectId)
-	if err != nil {
-		return err
-	}
-	defer fs.Close()
-
-	err = fs.RemoveAll(filepath.Join(util.WorkspaceSdksDir, blob.Name))
-	if err != nil {
-		return fmt.Errorf("cannot undo SDK %q installation: %w", sdkMount.Name, err)
-	}
-
-	return nil
-}
-
-func (m *WorkspaceManager) doLinkSdk(task *state.Task, tomb *tomb.Tomb) error {
-	project, workspace, err := ProjectAndWorkspace(task)
-	if err != nil {
-		return err
-	}
-
-	blob, err := sdkData(task)
-	if err != nil {
-		return err
-	}
-
-	st := task.State()
-	st.Lock()
-	defer st.Unlock()
-
-	/* Read a sequence record for the SDK (if any) */
-	props, err := m.backend.GetWorkspace(workspace, project.ProjectId)
-	if err != nil {
-		return err
-	}
-
-	var sequence = make(map[string][]*SdkSequenceRecord, 0)
-	if sdks, ok := props.Config["user.workspace.sdk"]; ok {
-		err = json.Unmarshal([]byte(sdks), &sequence)
-		if err != nil {
-			return err
-		}
-	}
-	sequence[blob.Name] = append(sequence[blob.Name], &SdkSequenceRecord{
-		blob.Channel, blob.Revision,
-	})
-
-	sequenceValue, err := json.Marshal(sequence)
-	if err != nil {
-		return err
-	}
-	/* Make a record in a LXD's key value storage to maintain
-	the sequence of the SDK's revisions */
-	err = m.backend.AddWorkspaceConfig(workspace, project.ProjectId,
-		&srv.WorkspaceConfigValue{
-			Name:  "user.workspace.sdk",
-			Value: string(sequenceValue),
-		})
-
-	if err != nil {
-		return err
-	}
-
-	/* Update the current link to point out to the newly installed SDK */
-	sdkPath := filepath.Join(util.WorkspaceSdksDir, blob.Name)
-
-	fs, err := m.backend.GetWorkspaceFs(workspace, project.ProjectId)
-	if err != nil {
-		return err
-	}
-	defer fs.Close()
-
-	err = fs.Symlink(filepath.Join(sdkPath, strconv.Itoa(int(blob.Revision))),
-		filepath.Join(sdkPath, "current"), true)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func (m *WorkspaceManager) undoLinkSdk(task *state.Task, tomb *tomb.Tomb) error {
-	project, workspace, err := ProjectAndWorkspace(task)
-	if err != nil {
-		return err
-	}
-
-	blob, err := sdkData(task)
-	if err != nil {
-		return err
-	}
-
-	st := task.State()
-	st.Lock()
-	defer st.Unlock()
-
-	/* Read a sequence record for the SDK (if any) */
-	props, err := m.backend.GetWorkspace(workspace, project.ProjectId)
-	if err != nil {
-		return err
-	}
-
-	var sequence = make(map[string][]*SdkSequenceRecord, 0)
-	if sdks, ok := props.Config["user.workspace.sdk"]; ok {
-		err = json.Unmarshal([]byte(sdks), &sequence)
-		if err != nil {
-			return err
-		}
-
-		/* Remove the latest sequence record */
-		seqLen := len(sequence[blob.Name])
-		if seqLen > 0 {
-			sequence[blob.Name] = sequence[blob.Name][:seqLen-1]
-		}
-
-		/* If no records in the SDK's sequence -- remove the SDK */
-		newSeqLen := len(sequence[blob.Name])
-		if newSeqLen == 0 {
-			delete(sequence, blob.Name)
-		}
-
-		if len(sequence) > 0 {
-			newSequence, err := json.Marshal(sequence)
-			if err != nil {
-				return err
-			}
-
-			/* Update the workspace config */
-			err = m.backend.AddWorkspaceConfig(workspace, project.ProjectId,
-				&srv.WorkspaceConfigValue{
-					Name:  "user.workspace.sdk",
-					Value: string(newSequence),
-				})
-			if err != nil {
-				return err
-			}
-		} else {
-			/* If no SDKs left in the sequence record, remove it fully */
-			err = m.backend.RemoveWorkspaceConfig(workspace, project.ProjectId,
-				"user.workspace.sdk")
-			if err != nil {
-				return nil
-			}
-		}
-		/* Update the 'current' link */
-		fs, err := m.backend.GetWorkspaceFs(workspace, project.ProjectId)
-		if err != nil {
-			return err
-		}
-		defer fs.Close()
-
-		if newSeqLen > 0 {
-			/* There is another revision available, shift the link to it */
-			err = fs.Symlink(strconv.Itoa(int(sequence[blob.Name][newSeqLen-1].Revision)),
-				filepath.Join(util.WorkspaceSdksDir, blob.Name, "current"), true)
-		} else {
-			/* It was the only revision, remove the link */
-			err = fs.Remove(filepath.Join(util.WorkspaceSdksDir, blob.Name, "current"))
-		}
-
-		return err
-	}
-
-	return nil
-}
-
-func sdkData(task *state.Task) (*store.SdkBlob, error) {
-	st := task.State()
-	st.Lock()
-	defer st.Unlock()
-
-	var retrieveId string
-	var blob store.SdkBlob
-
-	err := task.Get("sdk-retrieve-task", &retrieveId)
-
-	if err != nil {
-		return nil, err
-	}
-
-	retrieve := task.State().Task(retrieveId)
-	if retrieve == nil {
-		return nil, fmt.Errorf("internal error: no corresponding retrieve-sdk task found")
-	}
-
-	if err = retrieve.Get("sdk-blob", &blob); err != nil {
-		return nil, err
-	}
-	return &blob, nil
-}
-
-func sdkBlobDevice(sdk *store.SdkBlob) srv.WorkspaceDevice {
-	filename := store.ToSdkFilename(sdk.Name, sdk.Revision)
-
-	/* Bind-mount the SDK to the workspace */
-	return srv.WorkspaceDevice{
-		Name: sdk.Name,
-		Properties: map[string]string{"type": "disk", "source": filename,
-			"path": filepath.Join("/root", filepath.Base(filename))},
-	}
-}
-
-func ProjectAndWorkspace(task *state.Task) (*projectstate.ProjectKey, string, error) {
-	st := task.State()
-	var project projectstate.ProjectKey
-	var name string
-
-	st.Lock()
-	err := task.Change().Get("project-key", &project)
-	st.Unlock()
-
-	if err != nil {
-		return nil, "", fmt.Errorf("cannot get project for task %q: %v", task.ID(), err)
-	}
-
-	st.Lock()
-	err = task.Change().Get("workspace", &name)
-	st.Unlock()
-
-	if err != nil {
-		return nil, "", fmt.Errorf("cannot get workspace for task %q: %v", task.ID(), err)
-	}
-
-	return &project, name, nil
 }
