+++ internal/workspacebackend/workspace_file.go
@@ -1,4 +1,4 @@
-package workspace
+package workspacebackend
 
 import (
 	"fmt"
